+++ js/ui/drop_down_button.js
@@ -17,7 +17,7 @@ import { isPlainObject, isDefined } from '../core/utils/type';
 import { ensureDefined } from '../core/utils/common';
 import Guid from '../core/guid';
 import { format as formatMessage } from '../localization/message';
-import { getElementWidth, getPopupWidth } from '../core/utils/drop_down';
+import { getElementWidth, getPopupWidth } from './drop_down_editor/drop_down';
 
 const DROP_DOWN_BUTTON_CLASS = 'dx-dropdownbutton';
 const DROP_DOWN_BUTTON_CONTENT = 'dx-dropdownbutton-content';
@@ -331,7 +331,7 @@ const DropDownButton = Widget.inherit({
                 show: { type: 'fade', duration: 0, from: 0, to: 1 },
                 hide: { type: 'fade', duration: 400, from: 1, to: 0 }
             },
-            width: getElementWidth.bind(null, this.$element()),
+            width: () => getElementWidth(this.$element()),
             height: 'auto',
             shading: false,
             position: {
@@ -411,7 +411,7 @@ const DropDownButton = Widget.inherit({
     _popupOptionChanged: function(args) {
         const options = Widget.getOptionsFromContainer(args);
 
-        this._setPopupOption('width', getElementWidth.bind(null, this.$element()));
+        this._setPopupOption('width', () => getElementWidth(this.$element()));
 
         const optionsKeys = Object.keys(options);
         if(optionsKeys.indexOf('width') !== -1 || optionsKeys.indexOf('height') !== -1) {
@@ -423,7 +423,7 @@ const DropDownButton = Widget.inherit({
         const popupWidth = getPopupWidth(this.option('dropDownOptions.width'));
 
         if(popupWidth === undefined) {
-            this._setPopupOption('width', getElementWidth.bind(null, this.$element()));
+            this._setPopupOption('width', () => getElementWidth(this.$element()));
         }
     },
